% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{Auth}
\alias{Auth}
\title{Auth Class Constructor}
\description{
Auth Class Constructor

Auth Class Constructor
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Auth-new}{\code{Auth$new()}}
\item \href{#method-Auth-token}{\code{Auth$token()}}
\item \href{#method-Auth-get_token}{\code{Auth$get_token()}}
\item \href{#method-Auth-clone}{\code{Auth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Auth-new"></a>}}
\if{latex}{\out{\hypertarget{method-Auth-new}{}}}
\subsection{Method \code{new()}}{
This function initializes a new instance of the `Auth` class with the specified parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Auth$new(user = NULL, password = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user}}{A character string representing the username for authentication.}

\item{\code{password}}{A character string representing the password for authentication.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the `Auth` class.

Retrieve a Token

This function retrieves a previously generated token.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Auth-token"></a>}}
\if{latex}{\out{\hypertarget{method-Auth-token}{}}}
\subsection{Method \code{token()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Auth$token()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character string representing the retrieved token.

Generate a Token

This function generates a unique token for authentication or other purposes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Auth-get_token"></a>}}
\if{latex}{\out{\hypertarget{method-Auth-get_token}{}}}
\subsection{Method \code{get_token()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Auth$get_token()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character string representing the generated token.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Auth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Auth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Auth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
