% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{Client}
\alias{Client}
\title{Client Class}
\description{
Client Class

Client Class
}
\seealso{
\code{\link{show_terms}} to read the Terms and conditions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Client-new}{\code{Client$new()}}
\item \href{#method-Client-token}{\code{Client$token()}}
\item \href{#method-Client-get_token}{\code{Client$get_token()}}
\item \href{#method-Client-send_request}{\code{Client$send_request()}}
\item \href{#method-Client-show_terms}{\code{Client$show_terms()}}
\item \href{#method-Client-accept_terms}{\code{Client$accept_terms()}}
\item \href{#method-Client-datasets}{\code{Client$datasets()}}
\item \href{#method-Client-search}{\code{Client$search()}}
\item \href{#method-Client-get_querytemplate}{\code{Client$get_querytemplate()}}
\item \href{#method-Client-generate_query_template}{\code{Client$generate_query_template()}}
\item \href{#method-Client-clone}{\code{Client$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-new"></a>}}
\if{latex}{\out{\hypertarget{method-Client-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$new(user, password, overwrite = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user}}{character. String representing the username for authentication.}

\item{\code{password}}{character. String representing the password for authentication.}

\item{\code{overwrite}}{logical. `FALSE` indicating whether to overwrite or not existing credentials.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the `Client` class.

Retrieve a Token

This function retrieves a previously generated token.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-token"></a>}}
\if{latex}{\out{\hypertarget{method-Client-token}{}}}
\subsection{Method \code{token()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$token()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character string representing the retrieved token.

Generate a Token

This function generates a unique token for authentication or other purposes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-get_token"></a>}}
\if{latex}{\out{\hypertarget{method-Client-get_token}{}}}
\subsection{Method \code{get_token()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$get_token()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character string representing the generated token.

Send a Request

This function sends a specified request to a service.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-send_request"></a>}}
\if{latex}{\out{\hypertarget{method-Client-send_request}{}}}
\subsection{Method \code{send_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$send_request(req, path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{A list or object representing the request to be sent.}

\item{\code{path}}{An optional parameter specifying the path where the response should be save at.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A response object or list containing the result of the sent request.

Show Terms and Conditions

This function displays the terms and conditions for the services.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-show_terms"></a>}}
\if{latex}{\out{\hypertarget{method-Client-show_terms}{}}}
\subsection{Method \code{show_terms()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$show_terms()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An HTML document containing the terms and conditions in a collapsible format.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-accept_terms"></a>}}
\if{latex}{\out{\hypertarget{method-Client-accept_terms}{}}}
\subsection{Method \code{accept_terms()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$accept_terms(term_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term_id}}{A character vector of term_ids that you wish to accept.
If missing current status is returned.
Use "\strong{accept_all}" if you want to accept all terms at once.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame reflecting the actual acceptance status for each term.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-datasets"></a>}}
\if{latex}{\out{\hypertarget{method-Client-datasets}{}}}
\subsection{Method \code{datasets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$datasets(pattern = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pattern}}{character. A string to filter dataset names by matching text. If NULL (default), it lists all datasets. This isn't a regex search, but a simple text match.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list. Containing datasets and associated information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-search"></a>}}
\if{latex}{\out{\hypertarget{method-Client-search}{}}}
\subsection{Method \code{search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$search(query, limit = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{Character. String representing the search query.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list or data frame containing the search results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-get_querytemplate"></a>}}
\if{latex}{\out{\hypertarget{method-Client-get_querytemplate}{}}}
\subsection{Method \code{get_querytemplate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$get_querytemplate(datasetId, to_json = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasetId}}{character. 'datasetId' of the specific dataset.}

\item{\code{to_json}}{logical. If `FALSE` return list, otherwise `toJSON(resp, pretty = TRUE, auto_unbox=TRUE)`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list or json file containing the raw query options.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-generate_query_template"></a>}}
\if{latex}{\out{\hypertarget{method-Client-generate_query_template}{}}}
\subsection{Method \code{generate_query_template()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$generate_query_template(datasetId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasetId}}{A numeric or character ID representing the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A JSON representing the generated query template.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Client-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
