% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accessors.R
\name{control}
\alias{control}
\alias{control<-}
\alias{batch}
\alias{batch<-}
\alias{activeDataset}
\alias{activeDataset<-}
\alias{qualityControl}
\alias{qualityControl<-}
\alias{pipeline}
\alias{pipeline<-}
\title{HDA Metadata Accessors}
\usage{
control(object)

control(object) <- value

batch(object)

batch(object) <- value

activeDataset(object)

activeDataset(object) <- value

qualityControl(object)

qualityControl(object) <- value

pipeline(object)

pipeline(object) <- value
}
\description{
Utility functions to access or set the value of slots of a \code{\link{HDAMetadata}} object.
}
