% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallColonyExclude.R
\name{excludeSmallColonies}
\alias{excludeSmallColonies}
\title{Exclusion of Small Colonies}
\usage{
excludeSmallColonies(
  HDA,
  what = "auto",
  exclude = "all",
  method = "otsu",
  manualThreshold = NULL
)
}
\arguments{
\item{HDA}{An object of class \code{\link{HDA}}.}

\item{what}{Which dataset of the \code{\link{HDA}} object should be thresholded? Defaults to "\code{RS}" (raw size).}

\item{method}{The thresholding method. Currently supported options are "\code{otsu}" or "\code{manual}".
For otsu thresholding, the threshold is automatically decided based on minimising the within-class variance of the log-transformed colony sizes.
Defaults to "\code{otsu}".}

\item{manualThreshold}{In the case that \code{method} is set to "\code{manual}", then the manually chosen threshold should be supplied here.}
}
\value{
\code{smallColonyExclude} will return an object of class \code{\link{HDA}}, with an additional dataset called "\code{SC}" (small colony).
This is a logical matrix, with code{FALSE} representing strains which have been excluded.
}
\description{
Extremely small colonies in the control dataset should be removed from all datasets, as these often represent strains which did not wake up.
No useful interaction information can be obtained from these strains, whilst their inclusion will disrupt subsequent normalisation processes.
}
\details{
For simplicity, technical replicates are treated together.
Thus, if the median size of the technical replicate is below the threshold, all techincal replicates are excluded.
For details on how to specify technical replicates, see \code{\link{HDALibrary}}.
In addition, there may be cases where there are multiple batches, and each batch has a different control dataset.
In these cases, each control will be analysed separately, and then colonies identified as absent will only be excluded from that particular batch.
}
