% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDA.R
\name{HDA}
\alias{HDA}
\title{High-density Array}
\usage{
HDA(
  files,
  directory = getwd(),
  imageAnalysisSoftware = "gitter",
  libraryFile,
  batches = 1
)
}
\description{
\code{HDA} is an class for the representation of data from high-throughput microbial screens such as synthetic genetic arrays or chemical screens.
}
