% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmFit.R
\name{lmFit}
\alias{lmFit}
\title{lmFit}
\usage{
lmFit(
  object,
  design,
  ndups = 1,
  block = NULL,
  correlation,
  weights = NULL,
  method = "ls",
  what = "auto",
  exclude = "all",
  ...
)
}
\description{
lmFit
}
